Moves mobile nav into it's own component
