+++ src/pages/components/nav/index.tsx
@@ -1,20 +1,9 @@
 import React from "react";
-import {
-  chakra,
-  Box,
-  Flex,
-  useColorModeValue,
-  HStack,
-  useDisclosure,
-  VStack,
-  IconButton,
-  CloseButton,
-  Avatar,
-} from "@chakra-ui/react";
-import { AiOutlineMenu } from "react-icons/ai";
+import { chakra, Flex, HStack, Avatar } from "@chakra-ui/react";
 import Link from "next/link";
 import { useTranslation } from "react-i18next";
 import styles from "../../../../styles/Home.module.css";
+import MobileNav from "./mobile-nav";
 
 /**
  * Implementation inspired from the developer dao website
@@ -22,12 +11,15 @@ import styles from "../../../../styles/Home.module.css";
  * source: https://github.com/Developer-DAO/developerdao.com
  *  */
 export default function Nav() {
-  const bg = useColorModeValue("white", "gray.800");
-  const mobileNav = useDisclosure();
   const { t } = useTranslation();
 
   return (
-    <chakra.nav borderBottom="1px solid" borderColor="#2C3539" backgroundColor="#2C3539" minWidth="100vw">
+    <chakra.nav
+      borderBottom="1px solid"
+      borderColor="#2C3539"
+      backgroundColor="#2C3539!important"
+      minWidth="100vw"
+    >
       <Flex
         alignItems="center"
         justifyContent="space-between"
@@ -78,47 +70,7 @@ export default function Nav() {
             </Link>
           </HStack>
 
-          <Box display={{ base: "inline-flex", md: "none" }}>
-            <IconButton
-              display={{ base: "flex", md: "none" }}
-              aria-label="Open menu"
-              fontSize="20px"
-              color={useColorModeValue("grey.800", "inherit")}
-              variant="ghost"
-              icon={<AiOutlineMenu />}
-              onClick={mobileNav.onOpen}
-            />
-
-            <VStack
-              pos="absolute"
-              top={0}
-              left={0}
-              right={0}
-              display={mobileNav.isOpen ? "flex" : "none"}
-              flexDirection="column"
-              pt={7}
-              pb={7}
-              m={0}
-              bg={bg}
-              spacing={3}
-              rounded="sm"
-              shadow="sm"
-              zIndex="99"
-            >
-              <CloseButton
-                aria-label="Close menu"
-                onClick={mobileNav.onClose}
-              />
-
-              <Link href="/" passHref>
-                <a>{t("searchButtonTitle")}</a>
-              </Link>
-
-              <Link href="/glossary" passHref>
-                <a>{t("glossaryButton")}</a>
-              </Link>
-            </VStack>
-          </Box>
+          <MobileNav />
         </HStack>
       </Flex>
     </chakra.nav>
